requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib
bcrypt==4.0.1
python-multipart
pyjwt
slowapi
pytest
httpx
email-validator